requirements.txt
# # langchain
# # PyPDF2
# # python-dotenv
# # openai
# # faiss-cpu
# # altair
# # qdrant_client
# # Flask


# langchain==0.0.184
# PyPDF2==3.0.1
# python-dotenv==1.0.0
# # streamlit==1.18.1
# openai==0.27.6
# faiss-cpu==1.7.4
# altair==4
# # tiktoken==0.4.0
# qdrant_client==1.6.9

# Flask==3.0.0

# # uncomment to use huggingface llms
# # huggingface-hub==0.14.1

# # uncomment to use instructor embeddings
# # InstructorEmbedding==1.0.1
# # sentence-transformers==2.2.2
#######################################################################
##############    New  libraries for the updated code  ###############


openai==0.27.6
Pillow==9.0.1
textwrap3==0.9.2
Flask==2.0.3
python-dotenv==1.0.0
langchain==0.0.184
qdrant-client==1.6.9
google-generativeai==0.4.0
IPython==8.22.2

